Traefik publish set to false for deployment
I'm using Nginx for a webserver and thought I disabled everything involving traefik but I'm having issues with kamal deploy checking for traefik and trying to bind to port 80. My next idea is to remove all containers on the vps and deploy with traefik and do my steps to use nginx over traefik, or perhaps see how simple it is to use traefik over nginx.