Bring back class, comment out background styles
Surprisingly removing .sorted-image-container didn't affect image really on desktop, but messed with images on mobile. The goal was to remove the background because it didnt mesh well (and very hard to center all images inside container, since all images themselves may not have the subject centered in frame). Plan is to only upload images with background removed. I am curious about automating that as I did watch a video discussing that type of servvice automated, but I believe she plans to use a phone app for background removal.