Switch fully to Importmaps with esbuild
Website seems to function correctly, testing with javascript and stimulus implemation.

Whole purpose of this change was to make future tinkiner easier by knowing if I was fully using importmaps or not and how to make changes.

Also to allow stimlus controllers to update on the fly as this was usually a big issue using them